Another day 20 part one refactor
