+++ Day20 - Pulse Propagation/Puzzle/Models/Modules/CommunicationModuleType.cs
@@ -1,4 +1,4 @@
-namespace AdventOfCode.Year2023.Day20.Puzzle.Models;
+namespace AdventOfCode.Year2023.Day20.Puzzle.Models.Modules;
 
 internal enum CommunicationModuleType
 {
