+++ Day20 - Pulse Propagation/Puzzle/Models/Modules/MachineModules.cs
@@ -2,23 +2,19 @@ namespace AdventOfCode.Year2023.Day20.Puzzle.Models.Modules;
 
 internal sealed class MachineModules
 {
-    private const string ButtonModuleName = "BUTTON";
-
     private readonly Dictionary<string, CommunicationModule> _modules;
 
-    public CommunicationModule ButtonModule { get; }
+    public CommunicationModule EntryModule { get; }
 
-    public MachineModules(IEnumerable<CommunicationModule> modules, string moduleConnectedToButtonName)
+    public MachineModules(IEnumerable<CommunicationModule> modules, string entryModuleName)
     {
         _modules = modules.ToDictionary(m => m.Name);
-        if (!_modules.TryGetValue(moduleConnectedToButtonName, out var moduleConnectedToButton))
+        if (!_modules.TryGetValue(entryModuleName, out var entryModule))
         {
-            throw new ArgumentException($"Module collection does not contain a module named '{moduleConnectedToButtonName}'.", nameof(moduleConnectedToButtonName));
+            throw new ArgumentException($"Module collection does not contain a module named '{entryModuleName}'.", nameof(entryModuleName));
         }
 
-        ButtonModule = new BroadcasterModule(ButtonModuleName);
-        moduleConnectedToButton.ConnectInput(ButtonModule);
-        _modules.Add(ButtonModuleName, ButtonModule);
+        EntryModule = entryModule;
     }
 
     public void Reset()
